refactor(hello-work-searcher): tanstack startから、next.jsへ移行 - bff使うなら、RSCでいいんじゃね？っと思ったため、 - graph-qlでbffを書くことを考えたが、それによって生じる複雑性と、next.jsへ移行することによる複雑性を天秤にかけた時、next.jsによる複雑性の方が許容できると判断
