--- apps/hello-work-job-searcher/.prettierignore
@@ -1,4 +1,0 @@
-**/build
-**/public
-pnpm-lock.yaml
-routeTree.gen.ts
