+++ apps/hello-work-job-searcher/package.json
@@ -1,32 +1,23 @@
 {
   "name": "hello-work-job-searcher",
+  "version": "0.1.0",
   "private": true,
-  "sideEffects": false,
-  "type": "module",
   "scripts": {
-    "dev": "vite dev",
-    "build": "vite build && tsc --noEmit",
-    "start": "node .output/server/index.mjs",
-    "deploy": "npm run build && wrangler deploy"
+    "dev": "next dev --turbopack",
+    "build": "next build",
+    "start": "next start",
+    "lint": "next lint"
   },
   "dependencies": {
     "@sho/models": "workspace:*",
-    "@tanstack/react-router": "^1.130.2",
-    "@tanstack/react-router-devtools": "^1.130.2",
-    "@tanstack/react-start": "^1.130.2",
-    "react": "^19.0.0",
-    "react-dom": "^19.0.0",
-    "zod": "^3.24.2"
+    "react": "19.1.0",
+    "react-dom": "19.1.0",
+    "next": "15.4.5"
   },
   "devDependencies": {
-    "@types/node": "^22.5.4",
-    "@types/react": "^19.0.8",
-    "@types/react-dom": "^19.0.3",
-    "@vitejs/plugin-react": "^4.6.0",
-    "autoprefixer": "^10.4.20",
-    "typescript": "^5.7.2",
-    "vite": "^7.0.0",
-    "vite-tsconfig-paths": "^5.1.4",
-    "wrangler": "^4.26.1"
+    "typescript": "^5",
+    "@types/node": "^20",
+    "@types/react": "^19",
+    "@types/react-dom": "^19"
   }
 }
