+++ apps/hello-work-job-searcher/src/app/page.tsx
@@ -1,0 +1,6 @@
+import { redirect } from "next/navigation";
+
+export default function Home() {
+  redirect("/jobs");
+  return null;
+}
