--- apps/hello-work-job-searcher/src/components/DefaultCatchBoundary.tsx
@@ -1,47 +1,0 @@
-import {
-  ErrorComponent,
-  Link,
-  rootRouteId,
-  useMatch,
-  useRouter,
-} from "@tanstack/react-router";
-import type { ErrorComponentProps } from "@tanstack/react-router";
-
-export function DefaultCatchBoundary({ error }: ErrorComponentProps) {
-  const router = useRouter();
-  const isRoot = useMatch({
-    strict: false,
-    select: (state) => state.id === rootRouteId,
-  });
-
-  console.error("DefaultCatchBoundary Error:", error);
-
-  return (
-    <div className="min-w-0 flex-1 p-4 flex flex-col items-center justify-center gap-6">
-      <ErrorComponent error={error} />
-      <div className="flex gap-2 items-center flex-wrap">
-        <button
-          type="button"
-          onClick={() => {
-            router.invalidate();
-          }}
-        >
-          Try Again
-        </button>
-        {isRoot ? (
-          <Link to="/">Home</Link>
-        ) : (
-          <Link
-            to="/"
-            onClick={(e) => {
-              e.preventDefault();
-              window.history.back();
-            }}
-          >
-            Go Back
-          </Link>
-        )}
-      </div>
-    </div>
-  );
-}
