--- apps/hello-work-job-searcher/src/router.tsx
@@ -1,20 +1,0 @@
-import { createRouter as createTanStackRouter } from "@tanstack/react-router";
-import { DefaultCatchBoundary } from "./components/DefaultCatchBoundary";
-import { routeTree } from "./routeTree.gen";
-
-export function createRouter() {
-  const router = createTanStackRouter({
-    routeTree,
-    defaultPreload: "intent",
-    defaultErrorComponent: DefaultCatchBoundary,
-    scrollRestoration: true,
-  });
-
-  return router;
-}
-
-declare module "@tanstack/react-router" {
-  interface Register {
-    router: ReturnType<typeof createRouter>;
-  }
-}
