--- apps/hello-work-job-searcher/src/routes/_pathlessLayout/_nested-layout/route-b.tsx
@@ -1,10 +1,0 @@
-import { createFileRoute } from "@tanstack/react-router";
-export const Route = createFileRoute("/_pathlessLayout/_nested-layout/route-b")(
-  {
-    component: LayoutBComponent,
-  },
-);
-
-function LayoutBComponent() {
-  return <div>I'm B!</div>;
-}
