--- apps/hello-work-job-searcher/src/routes/index.tsx
@@ -1,12 +1,0 @@
-import { createFileRoute } from "@tanstack/react-router";
-export const Route = createFileRoute("/")({
-  component: Home,
-});
-
-function Home() {
-  return (
-    <div className="p-2">
-      <h3>Welcome Home!!!</h3>
-    </div>
-  );
-}
