--- apps/hello-work-job-searcher/vite.config.ts
@@ -1,17 +1,0 @@
-import { tanstackStart } from "@tanstack/react-start/plugin/vite";
-import viteReact from "@vitejs/plugin-react";
-import { defineConfig } from "vite";
-import tsConfigPaths from "vite-tsconfig-paths";
-
-export default defineConfig({
-  server: {
-    port: 3000,
-  },
-  plugins: [
-    tsConfigPaths({
-      projects: ["./tsconfig.json"],
-    }),
-    tanstackStart({ customViteReactPlugin: true, target: "cloudflare-module" }),
-    viteReact(),
-  ],
-});
