--- apps/hello-work-job-searcher/wrangler.jsonc
@@ -1,13 +1,0 @@
-{
-  "$schema": "node_modules/wrangler/config-schema.json",
-  "name": "hello-work-job-searcher",
-  "main": ".output/server/index.mjs",
-  "compatibility_date": "2025-07-28",
-  "compatibility_flags": ["nodejs_compat"],
-  "assets": {
-    "directory": ".output/public"
-  },
-  "observability": {
-    "enabled": true
-  }
-}
